Refactor numOfWindows method to return value of size_t type

+++ header/Utility/TextSplitter.h
@@ -9,7 +9,7 @@ class TextSplitter {
 private:
     static const int numOfThreads = 32;
 
-    int numOfWindows(long long int textLength, int patternLength);
+    size_t numOfWindows(long long int textLength, int patternLength);
     std::pair<int, int> lenOfChain(int windows);
 public:
     std::vector<std::pair<size_t, size_t>> splitText(long long int textLength, int patternLength);
